requirements.txt
pyatv
pytz
requests
python-dateutil
trakt.py
PyYAML
lxml
protobuf